Room descriptions, items, npc table of responses
did all of the basic room descriptions, item descriptions, and
supporter/scenery descriptions, so now the layout of my game is much
clearer and easier to build puzzles around. Set  up the first puzzle
(to ask luigi about book, and then having to retrieve water for him to
read book), but I have to work out the kinks still because it won’t
program. Also set up the npc table of responses for luigi, as well as
making the code for the ouija book. Once I know how to have items with
the same name in my game, I’ll be able to finish my other puzzle
(creating a mirror image), which will be given hints through the npc
table of responses.