Finalization of first puzzle
To make sure the first puzzle worked, I had to take out the NPC table
of responses. This was also because I plan on having two NPCs, and I
think just having instead commands will be simpler when differentiating
between/setting up the two character’s responses. Made player
description, added in sited sources (still more to put in though).
Deleted the code from ragnorak as it was causing problems, so i took
the code for books/pages from the inform recipe and that worked pretty
well. This is for the set up of the next puzzle. Made a to-do list so i
stay organized and know what to do next.