Working on puzzle #2
started the set up for the second puzzle, moved around some of the
items, too. Also, going to change the whole book. It was going to be
instructions and general warning for how to use a ouija board, but
that’s kind of boring, so I’m going to make it a diary of someone who
used a ouija board. The book is the clue for the 2nd puzzle.