Almost done!!
worked on so much last night and today. changed the title, added in
descriptions, changing descriptions, added items, created all the
different endings to the game (this took the most time, had to think
about what the player would do to try to escape), added more dialogue,
and fixed some bugs w/ the canteen and opening the ouija board. just
need to finish beta testing and finalizations. Need to add in comments
to code.